DocumentRepository.java: findAll(), count(); DocumentRepositoryAdapter.java: count() implemented, findAll() returning PageResult<>, DocumentRepositoryAdapterTest.java: checkFindAll() adjusted, checkDeleteAll() adjusted, checkCount() added

+++ src/main/java/pl/jakubholik90/domain/common/PageResult.java
@@ -6,7 +6,7 @@ public record PageResult<T>(
         List<T> content,
         int page,
         int size,
-        long totalElements,
+        int totalElements,
         int totalPages) {
 
     public boolean hasNext() {
